Fixes updating members with the wrong pool ID
On queens, if you update a member using the wrong pool ID, the user
will get the correct error response, but the load balancer will
remain locked. This patch corrects that by putting the pool
validation in the correct sequence.
This is already fixed in Rocky and newer as part of the provider
driver feature.

Change-Id: Id88bd2363d4cb0213096997146d0a79454289978
Story: 2007609
Task: 39599