Fix amphora image build
DIB dropped Python 2 support in its version 3.0.0 major release and as
so we had to pin it to 2.30.0 in stable/stein and stable/train branches.
However, upstream get-pip.sh recently introduced a regression where it
does not work well with Python <= 3.5. Sadly, DIB 2.30.0 points to
master get-pip.sh [1] and DIB is branchless so the fix cannot be
backported and we are in a deadlock scenario where Python <= 3.5 amphora
images can't be built.

To work-around this issue on Stein and Train releases, we have to make a
compromise: either build amphora images using master DIB and Python
>=3.6 only or be stuck (unless someone else can come up with another
approach, all ears). Ussuri and newer releases are not impacted by this
issue as they are Python >=3.6 only releases.

Additionally, in the process of fixing Stein CI jobs we noticed that
octavia-v1-dsvm-scenario jobs were still running on Ubuntu Xenial
(Python 3.5!) but should have been on Bionic. The nodeset was updated to
Bionic-based but yet they continued to fail. For unknown reasons, Zuul
didn't collect CI log files so we couldn't troubleshoot the issue
further (infra team was pinged on IRC, no response). Given that Stein is
in extended maintenance and these jobs are neutron-lbaas (EOLd), we have
reasons to believe there is no one interested in maintaining these jobs
to the level worth keeping them as voting.

[1] https://review.opendev.org/c/openstack/diskimage-builder/+/772254

Change-Id: I13da9a49ace803e957dc52e4882b78141bc7bef3